refactor: Move binding adapter static methods to Util package
