+++ app/src/main/java/com/kappzzang/jeongsan/ui/addexpense/AddExpenseActivity.kt
@@ -1,13 +1,9 @@
 package com.kappzzang.jeongsan.ui.addexpense
 
 import android.os.Bundle
-import android.widget.EditText
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.widget.addTextChangedListener
 import androidx.databinding.BindingAdapter
-import androidx.databinding.InverseBindingAdapter
-import androidx.databinding.InverseBindingListener
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.kappzzang.jeongsan.databinding.ActivityAddExpenseBinding
@@ -17,52 +13,6 @@ import kotlinx.coroutines.flow.StateFlow
 fun attachList(recyclerView: RecyclerView, items: StateFlow<List<ExpenseItemInput>>?) {
     items?.let {
         (recyclerView.adapter as? ExpenseItemListAdapter)?.submitList(it.value)
-    }
-}
-
-object BindingAdapter {
-    @BindingAdapter("stringText")
-    @JvmStatic
-    fun setTextValue(view: EditText, value: String?) {
-        view.setText(value?:"")
-    }
-
-    @InverseBindingAdapter(attribute = "stringText")
-    @JvmStatic
-    fun getTextValue(view: EditText): String? {
-        val parsedString = view.text.toString()
-        if (parsedString.isBlank()) {
-            return null
-        }
-        return parsedString
-    }
-
-    @BindingAdapter("integerText")
-    @JvmStatic
-    fun setIntegerValue(view: EditText, value: Int?) {
-        view.setText(value?.toString()?:"")
-    }
-
-    @InverseBindingAdapter(attribute = "integerText")
-    @JvmStatic
-    fun getIntegerValue(view: EditText): Int? {
-        return view.text.toString().toIntOrNull()
-    }
-
-    @BindingAdapter("integerTextAttrChanged")
-    @JvmStatic
-    fun setIntegerTextListener(view: EditText, listener: InverseBindingListener) {
-        view.addTextChangedListener {
-            listener.onChange()
-        }
-    }
-
-    @BindingAdapter("stringTextAttrChanged")
-    @JvmStatic
-    fun setStringTextListener(view: EditText, listener: InverseBindingListener) {
-        view.addTextChangedListener {
-            listener.onChange()
-        }
     }
 }
